fix ChaosRelic & Blights obtain
